Add data version to patch header

+++ lib/pc_api.h
@@ -174,7 +174,7 @@ PCSTATS;
 */
 
 #define PCPATCH_COMMON \
-	int type; \
+	uint16_t type; \
 	int8_t readonly; \
 	const PCSCHEMA *schema; \
 	uint32_t npoints;  \
